refactor: move credential tests into req4att test file
